requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.9.1
Pillow==10.2.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
